add newsletter editor and make trait file for Nodes Controller and Nodes Sources Controller.
