+++ themes/Rozier/Controllers/NodesController.php
@@ -39,6 +39,7 @@ use RZ\Roadiz\Core\ListManagers\EntityListManager;
 use RZ\Roadiz\CMS\Forms\SeparatorType;
 
 use Themes\Rozier\RozierApp;
+use Themes\Rozier\Traits\NodesTrait;
 
 use RZ\Roadiz\Core\Exceptions\EntityAlreadyExistsException;
 
@@ -54,6 +55,8 @@ use Symfony\Component\Validator\Constraints\NotBlank;
  */
 class NodesController extends RozierApp
 {
+    use NodesTrait;
+
     /**
      * List every nodes.
      *
@@ -542,18 +545,6 @@ class NodesController extends RozierApp
                     $response->prepare($request);
 
                     return $response->send();
-                } catch (\Exception $e) {
-                    $this->publishErrorMessage($request, $e->getMessage());
-
-                    $response = new RedirectResponse(
-                        $this->getService('urlGenerator')->generate(
-                            'nodesAddChildPage',
-                            array('nodeId' => $nodeId, 'translationId' => $translationId)
-                        )
-                    );
-                    $response->prepare($request);
-
-                    return $response->send();
                 }
             }
 
@@ -745,490 +736,5 @@ class NodesController extends RozierApp
         } else {
             return $this->throw404();
         }
-    }
-
-    /**
-     * @param array                              $data
-     * @param RZ\Roadiz\Core\Entities\NodeType    $type
-     * @param RZ\Roadiz\Core\Entities\Translation $translation
-     *
-     * @return RZ\Roadiz\Core\Entities\Node
-     */
-    protected function createNode($data, NodeType $type, Translation $translation)
-    {
-        if ($this->urlAliasExists(StringHandler::slugify($data['nodeName']))) {
-            $msg = $this->getTranslator()->trans(
-                'node.%name%.no_creation.urlAlias.alreadyExists',
-                array('%name%'=>$data['nodeName'])
-            );
-
-            throw new EntityAlreadyExistsException($msg, 1);
-        }
-        if ($this->nodeNameExists(StringHandler::slugify($data['nodeName']))) {
-            $msg = $this->getTranslator()->trans(
-                'node.%name%.no_creation.already_exists',
-                array('%name%'=>$data['nodeName'])
-            );
-
-            throw new EntityAlreadyExistsException($msg, 1);
-        }
-
-        try {
-            $node = new Node($type);
-            $node->setNodeName($data['nodeName']);
-            $this->getService('em')->persist($node);
-
-            $sourceClass = "GeneratedNodeSources\\".$type->getSourceEntityClassName();
-            $source = new $sourceClass($node, $translation);
-            $source->setTitle($data['nodeName']);
-
-            $this->getService('em')->persist($source);
-            $this->getService('em')->flush();
-
-            return $node;
-        } catch (\Exception $e) {
-            $msg = $this->getTranslator()->trans(
-                'node.%name%.noCreation.alreadyExists',
-                array('%name%'=>$node->getNodeName())
-            );
-            throw new EntityAlreadyExistsException($msg, 1);
-        }
-    }
-
-    /**
-     * @param array       $data
-     * @param Node        $parentNode
-     * @param Translation $translation
-     *
-     * @return RZ\Roadiz\Core\Entities\Node
-     */
-    protected function createChildNode($data, Node $parentNode = null, Translation $translation = null)
-    {
-        if ($this->urlAliasExists(StringHandler::slugify($data['nodeName']))) {
-            $msg = $this->getTranslator()->trans(
-                'node.%name%.no_creation.url_alias.already_exists',
-                array('%name%'=>$data['nodeName'])
-            );
-
-            throw new EntityAlreadyExistsException($msg, 1);
-        }
-        if ($this->nodeNameExists(StringHandler::slugify($data['nodeName']))) {
-            $msg = $this->getTranslator()->trans(
-                'node.%name%.no_creation.already_exists',
-                array('%name%'=>$data['nodeName'])
-            );
-
-            throw new EntityAlreadyExistsException($msg, 1);
-        }
-        $type = null;
-
-        if (!empty($data['nodeTypeId'])) {
-            $type = $this->getService('em')
-                        ->find(
-                            'RZ\Roadiz\Core\Entities\NodeType',
-                            (int) $data['nodeTypeId']
-                        );
-        }
-        if (null === $type) {
-            throw new \Exception("Cannot create a node without a valid node-type", 1);
-        }
-        if (null !== $parentNode && $data['parentId'] != $parentNode->getId()) {
-            throw new \Exception("Requested parent node does not match form values", 1);
-        }
-
-        $node = new Node($type);
-        $node->setParent($parentNode);
-        $node->setNodeName($data['nodeName']);
-        $this->getService('em')->persist($node);
-
-        $sourceClass = "GeneratedNodeSources\\".$type->getSourceEntityClassName();
-        $source = new $sourceClass($node, $translation);
-        $source->setTitle($data['nodeName']);
-        $this->getService('em')->persist($source);
-        $this->getService('em')->flush();
-
-        return $node;
-    }
-
-    /**
-     * @param string $name
-     *
-     * @return boolean
-     */
-    protected function urlAliasExists($name)
-    {
-        return (boolean) $this->getService('em')
-            ->getRepository('RZ\Roadiz\Core\Entities\UrlAlias')
-            ->exists($name);
-    }
-
-    /**
-     * @param string $name
-     *
-     * @return boolean
-     */
-    protected function nodeNameExists($name)
-    {
-        return (boolean) $this->getService('em')
-            ->getRepository('RZ\Roadiz\Core\Entities\Node')
-            ->exists($name);
-    }
-
-    /**
-     * Edit node base parameters.
-     *
-     * @param array                       $data
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     */
-    protected function editNode($data, Node $node)
-    {
-        $testingNodeName = StringHandler::slugify($data['nodeName']);
-        if ($testingNodeName != $node->getNodeName() &&
-                ($this->nodeNameExists($testingNodeName) ||
-                $this->urlAliasExists($testingNodeName))) {
-            $msg = $this->getTranslator()->trans('node.%name%.noUpdate.alreadyExists', array('%name%'=>$data['nodeName']));
-            throw new EntityAlreadyExistsException($msg, 1);
-        }
-        foreach ($data as $key => $value) {
-            if ($key == 'home' &&
-                true === (boolean) $value) {
-                $node->getHandler()->makeHome();
-            } else {
-                $setter = 'set'.ucwords($key);
-                $node->$setter( $value );
-            }
-        }
-
-        $this->getService('em')->flush();
-    }
-
-    /**
-     * @param array $data
-     * @param Node  $node
-     */
-    public function addStackType($data, Node $node)
-    {
-        if ($data['nodeId'] == $node->getId() &&
-            !empty($data['nodeTypeId'])) {
-            $nodeType = $this->getService('em')
-                 ->find('RZ\Roadiz\Core\Entities\NodeType', (int) $data['nodeTypeId']);
-
-            if (null !== $nodeType) {
-                $node->addStackType($nodeType);
-                $this->getService('em')->flush();
-
-                return $nodeType;
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Link a node with a tag.
-     *
-     * @param array                       $data
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     *
-     * @return RZ\Roadiz\Core\Entities\Tag $linkedTag
-     */
-    protected function addNodeTag($data, Node $node)
-    {
-        if (!empty($data['tagPaths'])) {
-            $paths = explode(',', $data['tagPaths']);
-            $paths = array_filter($paths);
-
-            foreach ($paths as $path) {
-                $tag = $this->getService('em')
-                            ->getRepository('RZ\Roadiz\Core\Entities\Tag')
-                            ->findOrCreateByPath($path);
-
-                $node->addTag($tag);
-            }
-        }
-
-        $this->getService('em')->flush();
-
-        return $tag;
-    }
-
-    /**
-     * @param array                       $data
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     * @param RZ\Roadiz\Core\Entities\Tag  $tag
-     *
-     * @return RZ\Roadiz\Core\Entities\Tag
-     */
-    protected function removeNodeTag($data, Node $node, Tag $tag)
-    {
-        if ($data['nodeId'] == $node->getId() &&
-            $data['tagId'] == $tag->getId()) {
-            $node->removeTag($tag);
-            $this->getService('em')->flush();
-
-            return $tag;
-        }
-    }
-
-    /**
-     * Create a new node-source for given translation.
-     *
-     * @param array                       $data
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     *
-     * @return void
-     */
-    protected function translateNode($data, Node $node)
-    {
-        $newTranslation = $this->getService('em')
-                ->find(
-                    'RZ\Roadiz\Core\Entities\Translation',
-                    (int) $data['translationId']
-                );
-
-        $baseSource = $node->getNodeSources()->first();
-
-        $source = clone $baseSource;
-
-        foreach ($source->getDocumentsByFields() as $document) {
-            $this->getService('em')->persist($document);
-        }
-        $source->setTranslation($newTranslation);
-        $source->setNode($node);
-
-        $this->getService('em')->persist($source);
-        $this->getService('em')->flush();
-    }
-
-    /**
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     *
-     * @return \Symfony\Component\Form\Form
-     */
-    protected function buildTranslateForm(Node $node)
-    {
-        $translations = $node->getHandler()->getUnavailableTranslations();
-        $choices = array();
-
-        foreach ($translations as $translation) {
-            $choices[$translation->getId()] = $translation->getName();
-        }
-
-        if ($translations !== null && count($choices) > 0) {
-            $builder = $this->getService('formFactory')
-                ->createBuilder('form')
-                ->add('nodeId', 'hidden', array(
-                    'data' => $node->getId(),
-                    'constraints' => array(
-                        new NotBlank()
-                    )
-                ))
-                ->add('translationId', 'choice', array(
-                    'label' => $this->getTranslator()->trans('translation'),
-                    'choices' => $choices,
-                    'required' => true
-                ));
-
-            return $builder->getForm();
-        } else {
-            return null;
-        }
-    }
-    /**
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     *
-     * @return \Symfony\Component\Form\Form
-     */
-    public function buildStackTypesForm(Node $node)
-    {
-        if ($node->isHidingChildren()) {
-            $defaults = array();
-
-            $builder = $this->getService('formFactory')
-                ->createBuilder('form', $defaults)
-                ->add('nodeId', 'hidden', array(
-                    'data'=>(int) $node->getId()
-                ))
-                ->add('nodeTypeId', new \RZ\Roadiz\CMS\Forms\NodeTypesType(), array(
-                    'label' => $this->getTranslator()->trans('nodeType'),
-                ));
-
-            return $builder->getForm();
-        } else {
-            return null;
-        }
-    }
-
-
-    /**
-     * @param RZ\Roadiz\Core\Entities\Node $parentNode
-     *
-     * @return \Symfony\Component\Form\Form
-     */
-    protected function buildAddChildForm(Node $parentNode = null)
-    {
-        $defaults = array();
-
-        $builder = $this->getService('formFactory')
-            ->createBuilder('form', $defaults)
-            ->add('nodeName', 'text', array(
-                'label' => $this->getTranslator()->trans('nodeName'),
-                'constraints' => array(
-                    new NotBlank()
-                )
-            ))
-            ->add('nodeTypeId', new \RZ\Roadiz\CMS\Forms\NodeTypesType(), array(
-                'label' => $this->getTranslator()->trans('nodeType'),
-            ));
-
-        if (null !== $parentNode) {
-            $builder->add('parentId', 'hidden', array(
-                'data'=>(int) $parentNode->getId(),
-                'constraints' => array(
-                    new NotBlank()
-                )
-            ));
-        }
-
-        return $builder->getForm();
-    }
-
-    /**
-     * @param RZ\Roadiz\Core\Entities\Node  $node
-     *
-     * @return \Symfony\Component\Form\Form
-     */
-    protected function buildEditForm(Node $node)
-    {
-        $defaults = array(
-            'nodeName' => $node->getNodeName(),
-            'home' => $node->isHome(),
-            'priority' => $node->getPriority(),
-            'dynamicNodeName' => $node->isDynamicNodeName(),
-        );
-        $builder = $this->getService('formFactory')
-            ->createBuilder('form', $defaults)
-            ->add(
-                'nodeName',
-                'text',
-                array(
-                    'label' => $this->getTranslator()->trans('nodeName'),
-                    'constraints' => array(new NotBlank())
-                )
-            )
-            ->add(
-                'priority',
-                'number',
-                array(
-                    'label' => $this->getTranslator()->trans('priority'),
-                    'constraints' => array(new NotBlank())
-                )
-            )
-            ->add(
-                'home',
-                'checkbox',
-                array(
-                    'label' => $this->getTranslator()->trans('node.isHome'),
-                    'required' => false,
-                    'attr' => array('class' => 'rz-boolean-checkbox')
-                )
-            )
-            ->add(
-                'dynamicNodeName',
-                'checkbox',
-                array(
-                    'label' => $this->getTranslator()->trans('node.dynamicNodeName'),
-                    'required' => false,
-                    'attr' => array('class' => 'rz-boolean-checkbox')
-                )
-            );
-
-        return $builder->getForm();
-    }
-
-    /**
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     *
-     * @return \Symfony\Component\Form\Form
-     */
-    protected function buildEditTagsForm(Node $node)
-    {
-        $defaults = array(
-            'nodeId' =>  $node->getId()
-        );
-        $builder = $this->getService('formFactory')
-                    ->createBuilder('form', $defaults)
-                    ->add('nodeId', 'hidden', array(
-                        'data' => $node->getId(),
-                        'constraints' => array(
-                            new NotBlank()
-                        )
-                    ))
-                    ->add('tagPaths', 'text', array(
-                        'label' => $this->getTranslator()->trans('list.tags.to_link'),
-                        'attr' => array('class' => 'rz-tag-autocomplete')
-                    ))
-                    ->add('separator_1', new SeparatorType(), array(
-                        'label' => $this->getTranslator()->trans('use.new_or_existing.tags_with_hierarchy'),
-                        'attr' => array('class' => 'form-help-static uk-alert uk-alert-large')
-                    ));
-
-        return $builder->getForm();
-    }
-
-    /**
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     *
-     * @return \Symfony\Component\Form\Form
-     */
-    protected function buildDeleteForm(Node $node)
-    {
-        $builder = $this->getService('formFactory')
-            ->createBuilder('form')
-            ->add('nodeId', 'hidden', array(
-                'data' => $node->getId(),
-                'constraints' => array(
-                    new NotBlank()
-                )
-            ));
-
-        return $builder->getForm();
-    }
-
-    /**
-     * @return \Symfony\Component\Form\Form
-     */
-    protected function buildEmptyTrashForm()
-    {
-        $builder = $this->getService('formFactory')
-            ->createBuilder('form');
-
-        return $builder->getForm();
-    }
-
-    /**
-     * @param RZ\Roadiz\Core\Entities\Node $node
-     * @param RZ\Roadiz\Core\Entities\Tag  $tag
-     *
-     * @return \Symfony\Component\Form\Form
-     */
-    protected function buildRemoveTagForm(Node $node, Tag $tag)
-    {
-        $builder = $this->getService('formFactory')
-            ->createBuilder('form')
-            ->add('nodeId', 'hidden', array(
-                'data' => $node->getId(),
-                'constraints' => array(
-                    new NotBlank()
-                )
-            ))
-            ->add('tagId', 'hidden', array(
-                'data' => $tag->getId(),
-                'constraints' => array(
-                    new NotBlank()
-                )
-            ));
-
-        return $builder->getForm();
     }
 }
